refactor: update recommended configs to use new file structure and remove deprecated files
