+++ packages/eslint-plugin-angular-template/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "exports": {
     ".": "./dist/index.esm.js",
-    "./configs/recommended": "./dist/configs/recommended.config.js"
+    "./configs/recommended": "./dist/configs/recommended.js"
   },
   "main": "dist/index.js",
   "typings": "typings/index.d.ts",
