+++ packages/eslint-plugin-angular-template/src/index.ts
@@ -1,4 +1,4 @@
 // @ts-ignore - Import JS file in TypeScript
-import recommended from './configs/recommended.config.js';
+import recommended from './configs/recommended.js';
 
 export { recommended };
