--- packages/eslint-plugin-angular/src/configs/base.config.js
@@ -1,26 +1,0 @@
-import typescriptParser from '@typescript-eslint/parser';
-import typescriptPlugin from '@typescript-eslint/eslint-plugin';
-import angularPlugin from '@angular-eslint/eslint-plugin';
-import { dirname } from 'path';
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-export default [
-  {
-    languageOptions: {
-      parser: typescriptParser,
-      parserOptions: {
-        ecmaVersion: 2020,
-        sourceType: 'module',
-        project: './tsconfig.json',
-        tsconfigRootDir: __dirname
-      }
-    },
-    plugins: {
-      '@typescript-eslint': typescriptPlugin,
-      '@angular-eslint': angularPlugin,
-    }
-  }
-];
