+++ packages/eslint-plugin-angular/src/configs/recommended.js
